* refactor importante de licencias y vacaciones (ahora unificados) * metodos que faltaban de employee * tests de eso * refactor en niveles de plan de carrera (y tests de eso) * entity's nuevas (para mapeos que faltaban)

+++ trunk/dopplereffect/dopplereffect-model/src/main/java/ar/edu/unq/dopplereffect/project/ProjectHelper.java
@@ -9,7 +9,7 @@ public class ProjectHelper {// NOPMD
     }
 
     public static long daysToHoursEffort(final int days) {
-        return days * HOURS_PER_DAY;
+        return (long) days * HOURS_PER_DAY;
     }
 
     public static int hoursEffortToDays(final Long effort) {
